+++ src/wirestate-lit/commands/on-command-controller.ts
@@ -5,7 +5,7 @@ import { CommandBus, CommandHandler, CommandType, CommandUnregister } from "@wir
 import { dbg } from "@/macroses/dbg.macro";
 import { prefix } from "@/macroses/prefix.macro";
 
-import { IocContextObject } from "../context/ioc-context";
+import { ContainerContext } from "../context/container-context";
 import { Optional } from "../types/general";
 
 /**
@@ -38,10 +38,10 @@ export class OnCommandController<D = unknown, R = unknown> implements ReactiveCo
     this.handler = handler;
 
     new ContextConsumer(host, {
-      context: IocContextObject,
+      context: ContainerContext,
       subscribe: true,
-      callback: (context) => {
-        this.bus = context.container.get(CommandBus);
+      callback: (container) => {
+        this.bus = container.get(CommandBus);
 
         if (host.isConnected) {
           this.reregister();